+++ VIKomet/SDK/Clients/ItemClient.cs
@@ -215,13 +215,13 @@ namespace VIKomet.SDK.Clients
             }
         }
 
-        public int Count()
+        public long Count()
         {
             HttpResponseMessage response = client.GetAsync("api/datastorage/item/count/quantity").Result;  // Blocking call!
             if (response.IsSuccessStatusCode)
             {
                 // Parse the response body. Blocking!
-                var r = response.Content.ReadAsAsync<int>().Result;
+                var r = response.Content.ReadAsAsync<long>().Result;
                 return r;
             }
             else
