+++ VIKomet/SDK/Clients/LogClient.cs
@@ -75,14 +75,14 @@ namespace VIKomet.SDK.Clients
                 throw ErrorParser(response.Content.ReadAsStringAsync().Result);
             }
         }
-          
-        public int Count()
+
+        public long Count()
         {
             HttpResponseMessage response = client.GetAsync("api/messaging/log/quantity").Result;  // Blocking call!
             if (response.IsSuccessStatusCode)
             {
                 // Parse the response body. Blocking!
-                var r = response.Content.ReadAsAsync<int>().Result;
+                var r = response.Content.ReadAsAsync<long>().Result;
                 return r;
             }
             else
