+++ VIKomet/SDK/Clients/OrderManagementClient.cs
@@ -388,13 +388,13 @@ namespace VIKomet.SDK.Clients
             }
         }
 
-        public int CountOrders()
+        public long CountOrders()
         {
             HttpResponseMessage response = client.GetAsync(string.Format("api/webstore/om/order/quantity")).Result;  // Blocking call!
             if (response.IsSuccessStatusCode)
             {
                 // Parse the response body. Blocking!
-                var r = response.Content.ReadAsAsync<int>().Result;
+                var r = response.Content.ReadAsAsync<long>().Result;
                 return r;
             }
             else
